+++ src/plugin/css.js
@@ -42,7 +42,7 @@
             module = replacement[1];
         }
 
-        sb.trigger('css:before-check', moduleName, module);
+        sb.trigger('*:before-check', moduleName, module, 'css');
         // If module exists or its a worker or node.js environment
         if (module || !sb.document) {
             callback(sb.initialized[moduleName] ? module : sb.require(moduleName));