+++ CHANGELOG.md
@@ -101,3 +101,4 @@
   - info of unused modules, suspicious globals and off-package modules paths
   - `"stats_auto": true` automatic statistics push
   - plugins depends
+  - image loader plugin